MAINT: refactor basis functions to layers._basis
